Added function generators for condition functions.
